Debugging
 - rsvd0 in the block group descriptor was a uint64 instead of uint32 as the spec said it should have been, which caused miss-alignment. I added the constructor-size-check thing from the alpha to prevent this from happening in the future.
 - blocks_in_group calculations were wrong, consolidated the correction into its own function.
 - accidentally used PAGE_SIZE instead of ODB_PAGESIZE

valgrind is clean, moving on to another test.

+++ include/oidadb-internal/odbfile.h
@@ -30,7 +30,7 @@ struct odb_block_group_desc {
 
 	// see ODB_SPEC_FLAG_GROUP_* constants
 	uint16_t flags;
-	uint64_t rsvd0;
+	uint32_t rsvd0;
 
 	struct odb_block blocks[1023];
